+++ src/pathtracer/ray.cu
@@ -1,7 +1,7 @@
 #include "ray.h"
 
 namespace CGL {
-    HOST_DEVICE Ray::Ray(const Vector3D o, const Vector3D d, int depth) {
+    HOST_DEVICE Ray::Ray(const Vector3D o, const Vector3D d, uint8_t depth) {
         // printf("Ray : %lf %lf %lf %lf %lf %lf \n", o.x, o.y, o.z, d.x, d.y, d.z);
         this->o = o;
         this->d = d;
@@ -10,7 +10,7 @@ namespace CGL {
         this->max_t = INFINITY;
         this->inv_d = 1/d;
     }
-    HOST_DEVICE Ray::Ray(const Vector3D o, const Vector3D d, double max_t, int depth) {
+    HOST_DEVICE Ray::Ray(const Vector3D o, const Vector3D d, float max_t, uint8_t depth) {
         this->inv_d = 1/d;
         this->max_t = max_t;
         Ray(o, d, depth);
